migrate junit 3.8.1 -> 4.12
