+++ src/test/java/nl/tudelft/model/GameTest.java
@@ -10,26 +10,22 @@ import static org.mockito.Mockito.when;
 
 import java.util.LinkedList;
 
-import junit.framework.Test;
-import junit.framework.TestSuite;
 import nl.tudelft.semgroup4.Resources;
 import nl.tudelft.semgroup4.collision.CollisionHandler;
+import org.junit.Test;
 import org.newdawn.slick.SlickException;
 import org.newdawn.slick.state.StateBasedGame;
 
 public class GameTest extends OpenGLTestCase {
 
-    public static Test suite() {
-        return new TestSuite(GameTest.class);
-    }
-
     /**
      * Test to see if the game constructor works.
      */
+    @Test
     public void testGame() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         assertFalse(game.equals(null));
@@ -39,10 +35,11 @@ public class GameTest extends OpenGLTestCase {
     /**
      * Test to check the current level getter.
      */
+    @Test
     public void testGetCurLevel() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         assertEquals(game.getCurLevel().getId(), 1);
@@ -51,11 +48,12 @@ public class GameTest extends OpenGLTestCase {
     /**
      * Test to check the player lives getter.
      */
+    @Test
     public void testGetPlayerLives() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
         when(mockedPlayer.getLives()).thenReturn(1);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         assertEquals(game.getPlayerLives(), 1);
@@ -64,10 +62,11 @@ public class GameTest extends OpenGLTestCase {
     /**
      * Test to check the reset players.
      */
+    @Test
     public void testResetPlayers() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         game.resetPlayers();
@@ -78,11 +77,12 @@ public class GameTest extends OpenGLTestCase {
      * Test to check the level reset.
      * @throws SlickException - Resources not found.
      */
+    @Test
     public void testLevelReset1() throws SlickException {
         Resources.init();
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         when(mockedPlayer.getLives()).thenReturn(1);
@@ -95,11 +95,12 @@ public class GameTest extends OpenGLTestCase {
      * Test to check the level reset.
      * @throws SlickException - Resources not found.
      */
+    @Test
     public void testLevelReset2() throws SlickException {
         Resources.init();
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         when(mockedPlayer.getLives()).thenReturn(0);
@@ -112,11 +113,12 @@ public class GameTest extends OpenGLTestCase {
      * Test to check the level reset.
      * @throws SlickException - Resources not found.
      */
+    @Test
     public void testNextLevel1() throws SlickException {
         Resources.init();
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         assertEquals(game.getCurLevel().getId(), 1);
@@ -128,11 +130,12 @@ public class GameTest extends OpenGLTestCase {
      * Test to check the level reset.
      * @throws SlickException - Resources not found.
      */
+    @Test
     public void testNextLevel2() throws SlickException {
         Resources.init();
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         assertEquals(game.getCurLevel().getId(), 1);
@@ -149,10 +152,11 @@ public class GameTest extends OpenGLTestCase {
     /**
      * Test to check the new Collisionhandler getter.
      */
+    @Test
     public void testGetNewCollisionHandler() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 0, 0);
         CollisionHandler<GameObject, GameObject> handler = null;
@@ -164,10 +168,11 @@ public class GameTest extends OpenGLTestCase {
     /**
      * Test to check the getter for container width.
      */
-    public void testGetContainerWidth() {
+    @Test
+    public void testGetContainerWidthw() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 1, 0);
         assertEquals(game.getContainerWidth(), 1);
@@ -176,10 +181,11 @@ public class GameTest extends OpenGLTestCase {
     /**
      * Test to check the getter for container height.
      */
+    @Test
     public void testGetContainerHeight() {
         StateBasedGame mockedSbg = mock(StateBasedGame.class);
         Player mockedPlayer = mock(Player.class);
-        LinkedList<Player> playerList = new LinkedList<Player>();
+        LinkedList<Player> playerList = new LinkedList<>();
         playerList.add(mockedPlayer);
         Game game = new Game(mockedSbg, playerList, 1, 1);
         assertEquals(game.getContainerWidth(), 1);
