--- src/test/java/nl/tudelft/model/LevelTest.java
@@ -1,5 +1,0 @@
-package nl.tudelft.model;
-
-public class LevelTest {
-
-}
